requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
Pillow>=10.0.0